requirements.txt
asgiref==3.6.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
Django
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.3.0
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.2.4
Werkzeug==2.2.3
zipp==3.15.0
